% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rf_perf.R
\name{calc_rf_perf}
\alias{calc_rf_perf}
\title{calculates rf performance}
\usage{
calc_rf_perf(data)
}
\arguments{
\item{data}{data with variables Y and Ypred}
}
\value{
tibble with meanError and bias
}
\description{
calculates rf performance
}
