% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rf_R2.R
\name{calc_rf_R2}
\alias{calc_rf_R2}
\title{calculates R2}
\usage{
calc_rf_R2(data, type = "Y")
}
\arguments{
\item{data}{data with variables Y and Y_pred if type="Y" or N_obse and N_pred if type="N"}

\item{type}{type of variables ("Y" or "N": defaults to "Y")}
}
\value{
R2
}
\description{
calculates R2
}
