% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_T_Q.R
\name{calc_T_Q}
\alias{calc_T_Q}
\title{Based on Qdata and Qdata back in time, calculate T_Q}
\usage{
calc_T_Q(time_obs, Qdata, Qdata_back)
}
\arguments{
\item{time_obs}{a time of observation in Qdata}

\item{Qdata}{Qdata}

\item{Qdata_back}{Qdata before period covered by Qdata}
}
\value{
the value of T_Q for the observation at time `time_obs` in `Qdata`
}
\description{
Based on Qdata and Qdata back in time, calculate T_Q
}
