% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Qdata.R
\name{complete_Qdata}
\alias{complete_Qdata}
\title{Completes Qdata with variables T_Q and S.}
\usage{
complete_Qdata(Qdata)
}
\value{
Qdata completed with variables T_Q and S
}
\description{
Completes Qdata with variables T_Q and S.
}
