% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Qdata_with_Vdata.R
\name{complete_Qdata_with_Vdata}
\alias{complete_Qdata_with_Vdata}
\title{Completes Qdata with variables Vmax,Vsum. => QVdata}
\usage{
complete_Qdata_with_Vdata(Qdata, Vdata)
}
\arguments{
\item{Qdata}{discharge data}

\item{Vdata}{wind data}
}
\value{
QVdata completed with variables Vmax,Vsum
}
\description{
Completes Qdata with variables Vmax,Vsum. => QVdata
}
