% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Wdata_with_Ddata.R
\name{complete_Wdata_with_Ddata}
\alias{complete_Wdata_with_Ddata}
\title{Interpolates Ddata to add descriptor variables to Wdata}
\usage{
complete_Wdata_with_Ddata(Wdata, Qdata, newvars = c("Q", "T_Q", "S", "rT_Q"))
}
\arguments{
\item{Ddata}{descriptive data}
}
\value{
Wdata completed with variables Q, T_Q, and S and rT_Q
}
\description{
Interpolates Ddata to add descriptor variables to Wdata
}
