% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Wdata_with_Qdata.R
\name{complete_Wdata_with_Qdata}
\alias{complete_Wdata_with_Qdata}
\title{Interpolates Qdata to add discharge variables to Wdata}
\usage{
complete_Wdata_with_Qdata(Wdata, Qdata)
}
\value{
Wdata completed with variables Q, T_Q, and S
}
\description{
Interpolates Qdata to add discharge variables to Wdata
}
