% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Qdata_back.R
\name{get_Qdata_back}
\alias{get_Qdata_back}
\title{Collects Qdata back in time (enough to get previous, high enough peak in discharge)}
\usage{
get_Qdata_back(Qdata)
}
\arguments{
\item{Qdata}{discharge data for which to look back in time}

\item{station}{gauging station}
}
\value{
Qdata_back: discharge data back in time. Length of the period is enough to ensure calculation of T_Q for Qdata.
}
\description{
Collects Qdata back in time (enough to get previous, high enough peak in discharge)
}
