% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Adata.R
\name{import_Adata}
\alias{import_Adata}
\title{Imports data regarding annotation times}
\usage{
import_Adata(path, site)
}
\arguments{
\item{path}{the path towards the annotation times file}

\item{site}{the name of the site considered}
}
\value{
Adata
}
\description{
Imports data regarding annotation times
}
