% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Qdata.R
\name{import_Qdata}
\alias{import_Qdata}
\title{Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)}
\usage{
import_Qdata(path, site = path)
}
\arguments{
\item{path}{the path towards the Qdata file}

\item{site}{the name of the site considered}
}
\value{
Qdata
}
\description{
Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)
}
\examples{
Qdata=import_Qdata("data-raw/Qdata/Qdatc_Ain.csv", site="Ain")
Qdata=import_Qdata("data-raw/Qdata/Qdatc_Allier.csv", site="Allier")
}
