% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Vdata.R
\name{import_Vdata}
\alias{import_Vdata}
\title{Imports wind data}
\usage{
import_Vdata(path, site = path)
}
\arguments{
\item{path}{the path towards the wind data file}
}
\value{
Vdata
}
\description{
Imports wind data
}
