% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Wdata.R
\name{import_Wdata}
\alias{import_Wdata}
\title{Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)}
\usage{
import_Wdata(path)
}
\arguments{
\item{path}{the path towards the events directory}
}
\value{
Wdata
}
\description{
Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)
}
