% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_Qdata.R
\name{interp_Qdata}
\alias{interp_Qdata}
\title{interpolate Qdata to a period}
\usage{
interp_Qdata(Qdata, timescale = "mins")
}
\arguments{
\item{Qdata}{data to interpolate}

\item{timescale}{time scale at which to interpolate data (defaults to minutes -"mins"-)}
}
\value{
summarised data
}
\description{
interpolate Qdata to a period
}
