% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rf.R
\name{predict_rf}
\alias{predict_rf}
\title{Runs model random forest object on new data}
\usage{
predict_rf(newdata, obj_rf, correction = c(1, 0))
}
\arguments{
\item{obj_rf}{random forest object}

\item{correction}{whether to correct covariate shift, in which case two coefficients (alpha,beta) should be provided. Defaults to (1,0) i.e. no correction.}

\item{data}{data with variables Q, rT_Q, S}
}
\value{
Wdata completed with column Ypred (log of hourly flux) and Npred (hourly flux)
}
\description{
Runs model random forest object on new data
}
