% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_1_wood_log_file.R
\name{read_1_wood_log_file}
\alias{read_1_wood_log_file}
\title{Reads one wood log file. To be used from import_Wdata()}
\usage{
read_1_wood_log_file(file)
}
\arguments{
\item{file}{the path to one log file}
}
\value{
the data from file
}
\description{
Reads one wood log file. To be used from import_Wdata()
}
\examples{
woody:::read_1_wood_log_file("data-raw/wood_data_Ain/event20071122/Wood Log 20071122 15h11m.txt")
woody:::read_1_wood_log_file("data-raw/wood_data_Ain/event20120101/Event_20120101.xlsx")
}
