% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{runs random forest model}
\usage{
run_rf(Wdatc_wt)
}
\arguments{
\item{Wdatc_wt}{data with wood data as waiting times and variables Q, rT_Q, S, Y}
}
\value{
random forest object
}
\description{
runs random forest model
}
