% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_hourly.R
\name{summarise_hourly}
\alias{summarise_hourly}
\title{summarise data hourly}
\usage{
summarise_hourly(data, type = "Wdatc", Adata = "notprovided")
}
\arguments{
\item{type}{of data either "Wdatc" or something else e.g. "Ddata"}
}
\value{
random forest object
}
\description{
summarise data hourly
}
