% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_Qdata.R
\name{summarise_Qdata}
\alias{summarise_Qdata}
\title{summarise Qdata to a period}
\usage{
summarise_Qdata(Qdata, sigma_coeffs)
}
\arguments{
\item{sigma_coeffs}{the coefficients providing sigma (residual standard deviation) with sigma=coef[1]*mu +coef[2]}

\item{data}{data to summarise}

\item{Adata}{the annotation times (if available)}
}
\value{
summarised data
}
\description{
summarise Qdata to a period
}
