% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_Wdata.R
\name{summarise_Wdata}
\alias{summarise_Wdata}
\title{summarise Wdata to a period}
\usage{
summarise_Wdata(Wdata, Adata = NULL)
}
\arguments{
\item{Wdata}{data to summarise}

\item{Adata}{the annotation times (if available)}
}
\value{
summarised data
}
\description{
summarise Wdata to a period
}
