% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wdata_as_waiting_times.R
\name{Wdata_as_waiting_times}
\alias{Wdata_as_waiting_times}
\title{Transform Wdata (1 line= 1 occurrence of wood) into Wdata_wt (1 line= time between two occurrences of wood)}
\usage{
Wdata_as_waiting_times(Wdata, maxWait = 1200)
}
\arguments{
\item{maxWait}{the maximum waiting time in seconds (defaults to 1200 seconds i.e. 20 minutes)}
}
\value{
Wdata completed with column W (waiting time between two wood occurrences)
}
\description{
Transform Wdata (1 line= 1 occurrence of wood) into Wdata_wt (1 line= time between two occurrences of wood)
}
