% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Qdata.R
\name{complete_Qdata}
\alias{complete_Qdata}
\title{Completes qtvar with variables T_Q and S. => Qdata}
\usage{
complete_Qdata(qtvar, qnorm = 1, site = "site")
}
\arguments{
\item{qtvar}{the discharge dataset (with varying time resolution).}

\item{qnorm}{standard value of discharge used to normalise variable Q. Defaults to 1 (no normalisation).}

\item{site}{the site name (defaults to "site")}
}
\value{
Qdata completed with variables T_Q, rT_Q and  S
}
\description{
Completes qtvar with variables T_Q and S. => Qdata
}
