% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Qdata.R
\name{get_Qdata}
\alias{get_Qdata}
\title{Collects Qdata back in time (enough to get previous, high enough peak in discharge)}
\usage{
get_Qdata(station, Wdata, extratime = 1200)
}
\arguments{
\item{station}{gauging station}

\item{Wdata}{wood data for which to collect Qdata}

\item{extratime}{time (in seconds) before and after wood occurrences for which also to collect Qdata}
}
\value{
Qdata: discharge data over period covered by Qdata, + extratime seconds before and extratime seconds after
}
\description{
Collects Qdata back in time (enough to get previous, high enough peak in discharge)
}
