% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Wdata.R
\name{import_Wdata}
\alias{import_Wdata}
\title{Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)}
\usage{
import_Wdata(path, site = path)
}
\arguments{
\item{path}{the path towards the events directory}

\item{site}{the name of the site considered}

\item{min_length}{the minimum length for a wood piece to be included in the dataset. Defaults to NA (no filtering).}

\item{sample_length}{whether to sample the wood pieces with no length provided, according to the proportion of pieces with Length>min_length. Defaults to FALSE.}
}
\value{
Wdata
}
\description{
Imports data regarding occurrence of wood from files under path (path>1 directory=1 event>all logs of the event)
}
