% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rf.R
\name{predict_rf}
\alias{predict_rf}
\title{Runs model random forest object on new data}
\usage{
predict_rf(obj_rf, newdata)
}
\arguments{
\item{obj_rf}{random forest object}

\item{data}{data with variables Q, rT_Q, S}
}
\value{
Wdata completed with column W (waiting time between two wood occurrences)
}
\description{
Runs model random forest object on new data
}
