% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{runs random forest model}
\usage{
run_rf(Wdata)
}
\arguments{
\item{Wdata}{data with variables Q, rT_Q, S, Y}
}
\value{
Wdata completed with column W (waiting time between two wood occurrences)
}
\description{
runs random forest model
}
