% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{runs random forest model}
\usage{
run_rf(Wdatcwt, pred_vars = c("Q", "S", "rT_Q"))
}
\arguments{
\item{pred_vars}{predictor variables by default c("Q","rT_Q","S")}

\item{Wdatc_wt}{data with wood data as waiting times}
}
\value{
random forest object
}
\description{
runs random forest model
}
