% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_period.R
\name{summarise_period}
\alias{summarise_period}
\title{summarise data to a period}
\usage{
summarise_period(data, type = "Wdata", period = "hour", Adata = NULL)
}
\arguments{
\item{data}{data to summarise}

\item{type}{of data either "Wdata" or "Qdata"}

\item{period}{period (can be "hour","day",...)}
}
\value{
summarised data
data=tib_WpQc_h$Wdata[[1]];type="Wdata";period="hour";Adata=tib_WpQc_h$Adata[[1]]
}
\description{
summarise data to a period
}
