% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_Qdata.R
\name{summarise_Qdata}
\alias{summarise_Qdata}
\title{summarise Qdata to a period}
\usage{
summarise_Qdata(Qdata, sigma)
}
\arguments{
\item{sigma}{the residual error in the estimates of Ypred}

\item{data}{data to summarise}

\item{Adata}{the annotation times (if available)}
}
\value{
summarised data
}
\description{
summarise Qdata to a period
}
